+++ pkgs/jnigen/lib/src/bindings/dart_generator.dart
@@ -469,6 +469,18 @@ class $name$typeParamsDef extends $superName {
       r'${node.binaryName}',
       \$p,
       _\$invokePointer,
+      [
+''');
+      final interfaceAsyncMethod = _InterfaceIfAsyncMethod(
+        resolver,
+        s,
+        implClassName: implClassName,
+      );
+      for (final method in node.methods) {
+        method.accept(interfaceAsyncMethod);
+      }
+      s.write('''
+      ],
     );
     final \$a = \$p.sendPort.nativePort; 
     _\$impls[\$a] = \$impl;
@@ -735,6 +747,18 @@ class _TypeGenerator extends TypeVisitor<String> {
   }
 }
 
+class _ImplReturnType extends _TypeGenerator {
+  _ImplReturnType(super.resolver);
+
+  @override
+  String visitPrimitiveType(PrimitiveType node) {
+    // Supporting both `Future<void>` for listener callbacks and `void` for
+    // blocking ones.
+    if (node.name == 'void') return '$_jni.FutureOr<void>';
+    return super.visitPrimitiveType(node);
+  }
+}
+
 class _TypeClass {
   final String name;
   final bool canBeConst;
@@ -745,22 +769,11 @@ class _TypeClass {
 /// Generates the type class.
 class _TypeClassGenerator extends TypeVisitor<_TypeClass> {
   final bool isConst;
-
-  /// Whether or not to return the equivalent boxed type class for primitives.
-  /// Only for interface implemetation.
-  final bool boxPrimitives;
-
-  /// Whether or not to find the correct type variable from the static map.
-  /// Only for interface implemetation.
-  final bool typeVarFromMap;
-
   final Resolver resolver;
 
   _TypeClassGenerator(
     this.resolver, {
     this.isConst = true,
-    this.boxPrimitives = false,
-    this.typeVarFromMap = false,
   });
 
   @override
@@ -768,8 +781,6 @@ class _TypeClassGenerator extends TypeVisitor<_TypeClass> {
     final innerTypeClass = node.type.accept(_TypeClassGenerator(
       resolver,
       isConst: false,
-      boxPrimitives: false,
-      typeVarFromMap: typeVarFromMap,
     ));
     final ifConst = innerTypeClass.canBeConst && isConst ? 'const ' : '';
     return _TypeClass(
@@ -788,8 +799,6 @@ class _TypeClassGenerator extends TypeVisitor<_TypeClass> {
     final definedTypeClasses = node.params.accept(_TypeClassGenerator(
       resolver,
       isConst: false,
-      boxPrimitives: false,
-      typeVarFromMap: typeVarFromMap,
     ));
 
     // Can be const if all the type parameters are defined and each of them are
@@ -831,15 +840,12 @@ class _TypeClassGenerator extends TypeVisitor<_TypeClass> {
   @override
   _TypeClass visitPrimitiveType(PrimitiveType node) {
     final ifConst = isConst ? 'const ' : '';
-    final name = boxPrimitives ? 'J${node.boxedName}' : 'j${node.name}';
+    final name = 'j${node.name}';
     return _TypeClass('$ifConst$_jni.${name}Type()', true);
   }
 
   @override
   _TypeClass visitTypeVar(TypeVar node) {
-    if (typeVarFromMap) {
-      return _TypeClass('_\$impls[\$p]!.${node.name}', false);
-    }
     return _TypeClass(node.name, false);
   }
 
@@ -853,6 +859,23 @@ class _TypeClassGenerator extends TypeVisitor<_TypeClass> {
   _TypeClass visitNonPrimitiveType(ReferredType node) {
     final ifConst = isConst ? 'const ' : '';
     return _TypeClass('$ifConst${_jObject}Type()', true);
+  }
+}
+
+class _ImplTypeClass extends _TypeClassGenerator {
+  _ImplTypeClass(super.resolver);
+
+  @override
+  _TypeClass visitTypeVar(TypeVar node) {
+    // Get the concrete type variable from the static map.
+    return _TypeClass('_\$impls[\$p]!.${node.name}', false);
+  }
+
+  @override
+  _TypeClass visitPrimitiveType(PrimitiveType node) {
+    final ifConst = isConst ? 'const ' : '';
+    final name = 'J${node.boxedName}';
+    return _TypeClass('$ifConst$_jni.${name}Type()', true);
   }
 }
 
@@ -1480,7 +1503,7 @@ class _AbstractImplMethod extends Visitor<Method, void> {
 
   @override
   void visit(Method node) {
-    final returnType = node.returnType.accept(_TypeGenerator(resolver));
+    final returnType = node.returnType.accept(_ImplReturnType(resolver));
     final name = node.finalName;
     final args = node.params.accept(_ParamDef(resolver)).join(', ');
     s.writeln('  $returnType $name($args);');
@@ -1496,7 +1519,7 @@ class _ConcreteImplClosureDef extends Visitor<Method, void> {
 
   @override
   void visit(Method node) {
-    final returnType = node.returnType.accept(_TypeGenerator(resolver));
+    final returnType = node.returnType.accept(_ImplReturnType(resolver));
     final name = node.finalName;
     final args = node.params.accept(_ParamDef(resolver)).join(', ');
     s.writeln('  final $returnType Function($args) _$name;');
@@ -1512,10 +1535,47 @@ class _ConcreteImplClosureCtorArg extends Visitor<Method, String> {
 
   @override
   String visit(Method node) {
-    final returnType = node.returnType.accept(_TypeGenerator(resolver));
+    final returnType = node.returnType.accept(_ImplReturnType(resolver));
     final name = node.finalName;
     final args = node.params.accept(_ParamDef(resolver)).join(', ');
     return 'required $returnType Function($args) $name,';
+  }
+}
+
+/// Dart async function type of an interface method.
+///
+/// For example: `Future<void> Function(int)`.
+class _InterfaceAsyncFunctionType extends Visitor<Method, String> {
+  final Resolver resolver;
+
+  _InterfaceAsyncFunctionType(this.resolver);
+
+  @override
+  String visit(Method node) {
+    assert(node.returnType.name == 'void');
+    const returnType = '$_core.Future<void>';
+    final args = node.params
+        .map((p) => p.type)
+        .accept(_TypeGenerator(resolver))
+        .join(', ');
+    return '$returnType Function($args)';
+  }
+}
+
+class _InterfaceNeverFunctionType extends Visitor<Method, String> {
+  final Resolver resolver;
+
+  _InterfaceNeverFunctionType(this.resolver);
+
+  @override
+  String visit(Method node) {
+    assert(node.returnType.name == 'void');
+    const returnType = '$_core.Never';
+    final args = node.params
+        .map((p) => p.type)
+        .accept(_TypeGenerator(resolver))
+        .join(', ');
+    return '$returnType Function($args)';
   }
 }
 
@@ -1528,7 +1588,7 @@ class _ConcreteImplMethod extends Visitor<Method, void> {
 
   @override
   void visit(Method node) {
-    final returnType = node.returnType.accept(_TypeGenerator(resolver));
+    final returnType = node.returnType.accept(_ImplReturnType(resolver));
     final name = node.finalName;
     final argsDef = node.params.accept(_ParamDef(resolver)).join(', ');
     final argsCall = node.params.map((param) => param.finalName).join(', ');
@@ -1568,6 +1628,46 @@ class _InterfaceMethodIf extends Visitor<Method, void> {
   }
 }
 
+/// The if statement within the async methods list to conditionally add methods.
+class _InterfaceIfAsyncMethod extends Visitor<Method, void> {
+  final Resolver resolver;
+  final StringSink s;
+  final String implClassName;
+
+  _InterfaceIfAsyncMethod(this.resolver, this.s, {required this.implClassName});
+
+  @override
+  void visit(Method node) {
+    if (node.returnType.name != 'void') {
+      return;
+    }
+    final signature = node.javaSig;
+    final name = node.finalName;
+    final asyncType = node.accept(_InterfaceAsyncFunctionType(resolver));
+    final neverType = node.accept(_InterfaceNeverFunctionType(resolver));
+    // If the implementation is using the callback passing style, look at the
+    // actual passed callback instead of the wrapper function. The wrapper is
+    // always going to return `FutureOr<void>`.
+    //
+    // If the callback simply throws its return type will be `Never`. As any
+    // function `R <F>` is a subtype of `Never <F>`, we should have a special
+    // case for it. It's not possible to throw from a listener function, so
+    // functions that only throw and hence have a `Never` return type, will be
+    // considered to be sync.
+    s.write('''
+        if ((
+              \$impl is _$implClassName &&
+              (\$impl as _$implClassName)._$name is $asyncType &&
+              (\$impl as _$implClassName)._$name is! $neverType
+            ) || (
+              \$impl.$name is $asyncType &&
+              \$impl.$name is! $neverType
+            ))
+          r'$signature',
+''');
+  }
+}
+
 /// Generates casting to the correct parameter type from the list of JObject
 /// arguments received from the call to the proxy class.
 class _InterfaceParamCast extends Visitor<Param, void> {
@@ -1583,13 +1683,7 @@ class _InterfaceParamCast extends Visitor<Param, void> {
 
   @override
   void visit(Param node) {
-    final typeClass = node.type
-        .accept(_TypeClassGenerator(
-          resolver,
-          boxPrimitives: true,
-          typeVarFromMap: true,
-        ))
-        .name;
+    final typeClass = node.type.accept(_ImplTypeClass(resolver)).name;
     s.write('\$a[$paramIndex].as($typeClass, releaseOriginal: true)');
     if (node.type.kind == Kind.primitive) {
       // Convert to Dart type.
@@ -1606,7 +1700,7 @@ class _InterfaceParamCast extends Visitor<Param, void> {
 ///
 /// Since Dart doesn't know that this global reference is still used, it might
 /// garbage collect it via `NativeFinalizer` thus making it invalid.
-/// This passes the ownership to Java using `setAsReleased`.
+/// This passes the ownership to Java using `toPointer()`.
 ///
 /// `toPointer` detaches the object from the `NativeFinalizer` and Java
 /// will clean up the global reference afterwards.